requirements.txt
# Core Dependencies
openai>=1.12.0
groq>=0.4.2
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20

# Document Processing
pypdf2>=3.0.1
pdfplumber>=0.10.3
pytesseract>=0.3.10
pillow>=10.2.0
pdf2image>=1.16.3

# Vector Store
chromadb>=0.4.22
faiss-cpu>=1.7.4

# Search & Retrieval
rank-bm25>=0.2.2
sentence-transformers>=2.2.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Web Framework
streamlit>=1.31.0
streamlit-option-menu>=0.3.6

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
tqdm>=4.66.0

# Evaluation
scikit-learn>=1.4.0

